--- src/main/java/no/ntnu/IDATA2304/Group1/server/communicator/ServerThread.java
@@ -1,47 +1,0 @@
-package no.ntnu.IDATA2304.Group1.server.communicator;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.OutputStream;
-import java.io.PrintWriter;
-import java.net.Socket;
-import org.json.JSONTokener;
-
-/**
- * This thread is responsible to handle client connection.
- *
- * @author www.codejava.net
- */
-
- public class ServerThread extends Thread {
-    private Socket socket;
- 
-    public ServerThread(Socket socket) {
-        this.socket = socket;
-    }
- 
-    public void run() {
-        try {
-            InputStream input = socket.getInputStream();
-            BufferedReader reader = new BufferedReader(new InputStreamReader(input));
- 
-            OutputStream output = socket.getOutputStream();
-            
-            String text;
-
-            do {
-                text = reader.readLine();
-                JSONTokener JSONObject = new JSONTokener(input);
-                System.out.println(JSONObject);
- 
-            } while (!text.equals("bye"));
- 
-            socket.close();
-        } catch (IOException ex) {
-            System.out.println("Server exception: " + ex.getMessage());
-            ex.printStackTrace();
-        }
-    }
-}
